+++ weatherApp/Model/MainModel.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Main: Codable {
+    let temp, feels_like, temp_min, temp_max: Double
+    let pressure, humidity: Int
+}
