+++ weatherApp/Model/WeatherModel.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Weather: Codable {
+    let description: String
+}
