+++ weatherApp/Model/WindModel.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Wind: Codable {
+    let speed: Double
+}
